requirements.txt
statsbombpy==1.13.1
pandas==2.2.2
apache-airflow==2.10.0
apache-airflow-providers-postgres==5.12.0
pytest==8.3.2
python-dotenv==1.0.1
psycopg2-binary==2.9.9
sqlalchemy==1.4.53
numpy==2.1.0
beautifulsoup4==4.12.3
fuzzywuzzy==0.18.0
python-Levenshtein==0.25.1